requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.12.2
black==22.12.0
Brotli==1.0.9
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
cryptography==36.0.1
cx-Oracle==8.3.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
Faker==9.8.0
h11==0.13.0
httprunner==4.3.5
idna==3.3
importlib-metadata==6.7.0
iniconfig==2.0.0
Jinja2==3.1.2
jmespath==0.9.5
loguru==0.4.1
MarkupSafe==2.1.3
modules==1.0.0
mypy-extensions==1.0.0
openpyxl==3.0.9
outcome==1.1.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.2.0
py==1.11.0
pycparser==2.21
pydantic==1.8.2
pymongo==4.0.1
pyOpenSSL==22.0.0
PySocks==1.7.1
pytest==7.4.0
pytest-base-url==2.0.0
pytest-html==3.2.0
pytest-metadata==3.0.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0.1
requests==2.31.0
selenium==4.1.2
sentry-sdk==0.14.4
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.5
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
typing-extensions==3.10.0.2
urllib3==1.26.8
urllib3-secure-extra==0.1.0
win32-setctime==1.1.0
wsproto==1.1.0
zipp==3.15.0

gmssl~=3.2.2
PyMySQL~=1.1.0